# Adjust for API differences in bench code