@JorgeBucaran 2018-01-27
@apostolos Maybe we need to tweak the benchmarks a bit more.

Classcat is meant to be used as:

```js
cc(["foo", "bar", "baz"]) // => foo bar baz
```

whereas classNames is like:

```js
cc("foo", "bar", "baz") // => foo
```