@apostolos 2018-01-28
I think that's what I'm doing already. You'll notice that before adding a benchmark function to the Suite, I first measure the args[] length:

- If args has more than one item I pass ```args``` to ```classcat``` as a single argument in order to be used as it should be, since ```args``` is **already an array**! For ```classNames``` I left it like before, since a classNames user would pass ```args.length``` number of arguments to the function.

- However, when args has a single argument, it is unfair to pass an array to ```classcat```. That's why I pass just the first element (args[0]). That way, we avoid penalizing classcat with an unnecessary iteration over a single-element array (in this case the "Objects" test).

You'll also notice that I pass different functions for each case to the suite to avoid measuring the args.length check.