@apostolos 2018-01-28
UPDATE:

I refactored the code a bit:
- A single object argument is passed for both ```classNames``` and ```classcat``` in the "Objects" test
- In multi-item tests, the args[] array:
  - is passed directly to ```classcat```
  - it is expanded as arguments with apply to ```classNames```

That way we avoid any potential object/array lookups when not needed and function calls are always how they are supposed to be used.

### New bench numbers (on a i7-6500U @ 2.50GHz):

```
Classcat – Strings × 8,213,876 ops/sec
classNames – Strings × 3,677,145 ops/sec
Fastest is Classcat – Strings

Classcat – Objects × 7,448,948 ops/sec
classNames – Objects × 3,681,655 ops/sec
Fastest is Classcat – Objects

Classcat – Strings & Objects × 5,119,115 ops/sec
classNames – Strings & Objects × 3,015,089 ops/sec
Fastest is Classcat – Strings & Objects

Classcat – Mixed × 2,983,494 ops/sec
classNames – Mixed × 2,228,106 ops/sec
Fastest is Classcat – Mixed

Classcat – Arrays × 2,550,737 ops/sec
classNames – Arrays × 910,095 ops/sec
Fastest is Classcat – Arrays
```