@JorgeBucaran 2018-01-28
@apostolos Okay, I finally a chance to have a look at this and you were totally right! I still think your code can be simplified a bit more, so I'll just take care of it. 😅 

I want to rewrite the bench this way instead:

```js
suite
  .add(`Classcat – ${fixed.description}`, () => cc(fixed.args))
  .add(`classNames – ${fixed.description}`, () => cx.apply({}, fixed.args))
```