@JorgeBucaran 2018-01-28
@apostolos I've updated the benchmarks and results and added them to https://github.com/JorgeBucaran/classcat/pull/16. Thanks! 👋